Replace RecaptchaForm with RecaptchaEnterpriseForm (#12954)
changelog: Internal, reCaptcha, Replace RecaptchaForm with RecaptchaEnterpriseForm

Fixes https://gsa-standard.atlassian-us-gov-mod.net/browse/LG-17458

- Make RecaptchaMockForm inherit from RecaptchaEnterpriseForm
- Remove RecaptchaForm
